--- a/LICENSE
+++ b/LICENSE
@@ -1,4 +1,4 @@
-Copyright 2019 Allen Institute for Cell Science
+Copyright 2020 Allen Institute for Cell Science
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met: