--- a/README.md
+++ b/README.md
@@ -37,12 +37,13 @@ img = AICSImage("my_file.tiff")
 img.data  # returns 6D STCZYX numpy array
 img.dims  # returns string "STCZYX"
 img.shape  # returns tuple of dimension sizes in STCZYX order
+img.get_image_data("CZYX", S=0, T=0)  # returns 4D CZYX numpy array
 
 # Get 6D STCZYX numpy array
 data = imread("my_file.tiff")
 ```
 
-### Delayed Image Slice Reading
+### Delayed Image Reading
 ```python
 from aicsimageio import AICSImage, imread_dask
 
@@ -52,7 +53,6 @@ img.dask_data  # returns 6D STCZYX dask array
 img.dims  # returns string "STCZYX"
 img.shape  # returns tuple of dimension sizes in STCZYX order
 img.size("STC")  # returns tuple of dimensions sizes for just STC
-img.get_image_data("CZYX", S=0, T=0)  # returns 4D CZYX numpy array
 img.get_image_dask_data("CZYX", S=0, T=0)  # returns 4D CZYX dask array
 
 # Read specified portion of dask array
@@ -65,6 +65,13 @@ lazy_s0t0 = lazy_data[0, 0, :]
 s0t0 = lazy_s0t0.compute()
 ```
 
+#### Quick Start Notes
+In short, if the word "dask" appears in the function or property name, the function
+utilizes delayed reading, if not, the underlying operation is backed by the image fully
+read into memory. I.E. `AICSImage.data` and `AICSImage.get_image_data` load the entire
+image into memory, and `AICSImage.dask_data` and `AICSImage.get_image_dask_data` do not
+load any image data until the user calls `compute` on the `dask.Array` object and only
+the requested chunk will be loaded into memory instead of the entire image.
 
 ### Speed up IO and Processing with Dask Clients and Clusters
 If you have already spun up a `distributed.Client` object in your Python process or