--- a/aicsimageio/tests/readers/test_ome_tiff_reader.py
+++ b/aicsimageio/tests/readers/test_ome_tiff_reader.py
@@ -14,8 +14,18 @@ from aicsimageio.readers.ome_tiff_reader import OmeTiffReader
     "filename, " "expected_shape, " "expected_dims, " "select_scene",
     [
         ("s_1_t_1_c_1_z_1.ome.tiff", (325, 475), "YX", 0),
-        ("s_1_t_1_c_10_z_1.ome.tiff", (10, 1736, 1776), "CYX", 0,),
-        ("s_3_t_1_c_3_z_5.ome.tiff", (3, 5, 3, 325, 475), "SZCYX", 0,),
+        (
+            "s_1_t_1_c_10_z_1.ome.tiff",
+            (10, 1736, 1776),
+            "CYX",
+            0,
+        ),
+        (
+            "s_3_t_1_c_3_z_5.ome.tiff",
+            (3, 5, 3, 325, 475),
+            "SZCYX",
+            0,
+        ),
         pytest.param(
             "example.txt",
             None,
@@ -33,7 +43,11 @@ from aicsimageio.readers.ome_tiff_reader import OmeTiffReader
     ],
 )
 def test_ome_tiff_reader(
-    resources_dir, filename, expected_shape, expected_dims, select_scene,
+    resources_dir,
+    filename,
+    expected_shape,
+    expected_dims,
+    select_scene,
 ):
     # Get file
     f = resources_dir / filename