--- a/docs/conf.py
+++ b/docs/conf.py
@@ -59,9 +59,9 @@ templates_path = ["_templates"]
 # You can specify multiple suffix as a list of string:
 #
 source_suffix = {
-        ".rst": "restructuredtext",
-        ".txt": "markdown",
-        ".md": "markdown",
+    ".rst": "restructuredtext",
+    ".txt": "markdown",
+    ".md": "markdown",
 }
 
 # The master toctree document.
@@ -69,7 +69,7 @@ master_doc = "index"
 
 # General information about the project.
 project = u"aicsimageio"
-copyright = u'2019, Allen Institute for Cell Science'
+copyright = u"2019, Allen Institute for Cell Science"
 author = u"Allen Institute for Cell Science"
 
 # The version info for the project you"re documenting, acts as replacement
@@ -134,15 +134,12 @@ latex_elements = {
     # The paper size ("letterpaper" or "a4paper").
     #
     # "papersize": "letterpaper",
-
     # The font size ("10pt", "11pt" or "12pt").
     #
     # "pointsize": "10pt",
-
     # Additional stuff for the LaTeX preamble.
     #
     # "preamble": "",
-
     # Latex figure (float) alignment
     #
     # "figure_align": "htbp",
@@ -152,9 +149,13 @@ latex_elements = {
 # (source start file, target name, title, author, documentclass
 # [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, "aicsimageio.tex",
-     u"aicsimageio Documentation",
-     u"Allen Institute for Cell Science", "manual"),
+    (
+        master_doc,
+        "aicsimageio.tex",
+        u"aicsimageio Documentation",
+        u"Allen Institute for Cell Science",
+        "manual",
+    ),
 ]
 
 
@@ -162,11 +163,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, "aicsimageio",
-     u"aicsimageio Documentation",
-     [author], 1)
-]
+man_pages = [(master_doc, "aicsimageio", u"aicsimageio Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------
@@ -175,10 +172,13 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, "aicsimageio",
-     u"aicsimageio Documentation",
-     author,
-     "aicsimageio",
-     "A Python library for reading and writing image data with specific support for handling bio-formats.",
-     "Image Processing, Computational Biology"),
+    (
+        master_doc,
+        "aicsimageio",
+        u"aicsimageio Documentation",
+        author,
+        "aicsimageio",
+        "A Python library for reading and writing image data with specific support for handling bio-formats.",
+        "Image Processing, Computational Biology",
+    ),
 ]