## Description
We have received enough feedback regarding the speed of getting back image chunks using `get_image_data` on `AICSImage` and `Reader` that it is time to bite the bullet and simply change to "dask" means delayed, anything else uses in memory data.

This PR makes it so that `get_image_data` on both the `AICSImage` object and `Reader` object uses the `data` property instead of routing to `get_image_dask_data` and calling `compute` on the result. This should reduce the time it takes for the `get_image_data` function to return in two scenarios:

1. when the user has `aicsimageio.use_dask(False)` / no backing `distributed` cluster
2. when the user had not previously preloaded the data with `AICSImage.data` as an empty statement

In short, timing results for current and for this branch are below:
**current:**
* duration of first `get_image_data` call: 9.15s
* duration of subsequent `get_image_data` calls: 9.35s

_No change in subsequent to first calls because it is always reading using the dask array_

**this branch:**
* duration of first `get_image_data` call: 7.42s
* duration of subsequent `get_image_data` calls: 82ms

_Change in subsequent calls because the image is loaded in full during the first call_

This is a weird semi-non-semi breaking change. This is not a breaking change in End User API -- this is a breaking change in the single case where a user was using `get_image_data` on an image that was too large for memory and now this will attempt to load the entire image into memory before returning to `numpy`. Because of this, this will get it's entirely own 3.3.0 release.

_Additional changes from updating black formatting_

## Pull request recommendations:
- [x] Name your pull request _your-development-type/short-description_. Ex: _feature/read-tiff-files_
- [x] Link to any relevant issue in the PR description. Ex: _Resolves [gh-12], adds tiff file format support_
- [x] Provide relevant tests for your feature or bug fix.
- [x] Provide or update documentation for any feature added by your pull request.

IPython log using current:
```
In [1]: from aicsimageio import AICSImage

In [2]: img = AICSImage("aicsimageio/tests/resources/3500001004_100X_20170623_6-Scene-2-P33-F05.ome.tiff")

In [3]: img.shape
Out[3]: (1, 1, 7, 65, 624, 924)

In [4]: %time img.get_image_data("ZYX", S=0, T=0, C=0).shape
CPU times: user 9.78 s, sys: 1.64 s, total: 11.4 s
Wall time: 9.15 s
Out[4]: (65, 624, 924)

In [5]: %time img.get_image_data("ZYX", S=0, T=0, C=1).shape
CPU times: user 10.3 s, sys: 1.45 s, total: 11.8 s
Wall time: 9.35 s
Out[5]: (65, 624, 924)
```

IPython log using this branch:
```
In [1]: from aicsimageio import AICSImage

In [2]: img = AICSImage("aicsimageio/tests/resources/3500001004_100X_20170623_6-Scene-2-P33-F05.ome.tiff")

In [3]: img.shape
Out[3]: (1, 1, 7, 65, 624, 924)

In [4]: %time img.get_image_data("ZYX", S=0, T=0, C=0).shape
CPU times: user 7.22 s, sys: 198 ms, total: 7.42 s
Wall time: 7.42 s
Out[4]: (65, 624, 924)

In [5]: %time img.get_image_data("ZYX", S=0, T=0, C=1).shape
CPU times: user 79 µs, sys: 0 ns, total: 79 µs
Wall time: 82.3 µs
Out[5]: (65, 624, 924)
```

Thanks for contributing!