@toloudis 2020-09-03
I like it.  Users of get_image_data should use it if they know their image fits in memory.  It's a convenience API on top of numpy operations. 
   
One thought: if I want to write a very smart reader that actually knows how to only read the requested data, how would I do it?