@toloudis 2020-09-03
> So very possible to do that on the `Reader` level because that _is_ the lower level access. But I don't think we would ever be able to do it with `AICSImage` because "one object to rule them all" fails in that specific scenario

IF get_image_data doesn't use .data, but instead just calls a similar get_image_data down at the reader level, then we can do the reads without caching but they will not have the memory requirement of loading the whole image.